Det Virker

+++ MapController/SimEnvironment/Collision.cs
@@ -64,10 +64,10 @@ namespace SimEnvironment
                 return false;
         }
         
-        public bool CheckCollison(int posx, int posy, int x1, int y1, int x2, int y2)
+        public bool CheckCollison(double posx, double posy, int x1, int y1, int x2, int y2)
         {
-            if (Check(posx, posy, x1, y1) == true)
-                return Check(posx, posy, x2, y2);
+            if (Check((int)posx, (int)posy, x1, y1) == true)
+                return Check((int)posx, (int)posy, x2, y2);
             else
                 return false;
         }
